HIR-59 refactor: 패키지 경로 수정
